requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI & ML
openai-whisper==20231117
openai==1.3.5
anthropic==0.7.1
google-generativeai==0.3.1

# Database
supabase==2.0.3
psycopg2-binary==2.9.9

# Frontend
streamlit==1.28.2

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Email Processing
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx>=0.24.0
pandas==2.1.3
numpy==1.26.2
